+++ core/RELEASE_NOTES.md
@@ -1,8 +1,13 @@
 # Release Notes
 
+## Version 3.3.2
+* Added reading of gaplist from a file to `playback.py`.
+* Minor improvements to `rsn_gaps.py`.
+
 ## Version 3.3.1
 * Modified `core.playback.gapListObj.dump()` routine to return JSON element when given `filename='JSON'` as an argument.
 * Added log file path flag to `rsn_gaps.py`.
+* Added gap logging output.
 
 ## Version 3.3.0
 * Added functionality to detect gaps in `rsn_gaps.py` which is built off the rest of the package.